bottom navigation added with screens
