+++ Quicktournament/package.json
@@ -15,13 +15,16 @@
     "@fortawesome/free-solid-svg-icons": "^6.1.2",
     "@fortawesome/react-native-fontawesome": "^0.3.0",
     "@react-navigation/drawer": "^6.5.7",
+    "@react-navigation/material-bottom-tabs": "^6.2.15",
     "@react-navigation/native": "^6.1.2",
+    "@react-navigation/stack": "^6.3.16",
     "@rneui/base": "^4.0.0-rc.7",
     "@rneui/themed": "^4.0.0-rc.7",
     "babel-plugin-module-resolver": "^5.0.0",
     "cocoapods": "^0.0.0",
     "date-fns": "^2.29.3",
     "event-target-shim": "^6.0.2",
+    "moment": "^2.29.4",
     "native-base": "^3.4.25",
     "react": "^18.0.0",
     "react-hook-form": "^7.42.0",
