--- Quicktournament/screens/Games.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-import { View, Text } from "react-native";
-
-const Games = () => {
-    return (
-        <View>
-            <Text>
-                ongoing games here
-            </Text>
-        </View>
-    )
-}
-
-export default Games;
