--- Quicktournament/screens/HomeScreen.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import { View, Text } from "react-native";
-
-const HomeScreen = () => {
-    return (
-        <View>
-            <Text>
-                Homepage
-            </Text>
-        </View>
-    )
-}
-export default HomeScreen;
