--- Quicktournament/screens/WaitingList.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-import PlayerList from "../components/players/PlayerList";
-
-const WaitingList = ({ route }) => {
-    return (
-        <PlayerList route={route} />
-    )
-}
-
-export default WaitingList;
