+++ Quicktournament/src/screens/Scores.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { View, Text } from "react-native";
+
+const Scores = () => {
+    return (
+        <View>
+            <Text>
+                Sijoitukset
+            </Text>
+        </View>
+    )
+}
+export default Scores;
