+++ Quicktournament/src/screens/WaitingList.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import PlayerList from "../components/players/PlayerList";
+
+const WaitingList = ({ route }) => {
+    return (
+        <PlayerList route={route} />
+    )
+}
+
+export default WaitingList;
