+++ Quicktournament/src/utils/utils.tsx
@@ -1,0 +1,5 @@
+import moment from "moment";
+
+export function currentDate() {
+    return moment().format('dddd MM yyyy')
+}
